requirements.txt
# Core Agent Framework
langchain==0.1.17
langgraph==0.0.38
langchain-openai==0.0.8
langchain-community==0.0.36

# Database & Vector Operations
psycopg2-binary==2.9.9
asyncpg==0.29.0
greenlet==3.2.4
sqlalchemy==2.0.25
alembic==1.13.1
pgvector==0.2.4

# Vector Processing (Cloud-Based)
openai==1.35.0
google.generativeai=0.8.5
numpy==1.24.3
faiss-cpu==1.7.4

# Web Framework & Frontend
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
streamlit==1.28.1

# Production Infrastructure
celery==5.3.4
redis==5.0.1
prometheus-client==0.19.0
structlog==23.2.0
langsmith==0.1.17

# Additional utilities
python-dotenv==1.0.0
httpx==0.25.2
requests==2.31.0
tiktoken==0.9.0

wikipedia==1.4.0
duckduckgo-search==7.3.0
arxiv==2.2.0